requirements.txt
pandas
geopandas
matplotlib
seaborn
statsmodels
openpyxl
requests
python-dotenv
scikit-learn
